Win Scene Fixing
I made a win scene for both levels. It is now possible to go from the main menu all the way to a win condition, and repeat. I've made a complete game cycle as of this update. I might also add some extra stuff like moving platforms and enemies depending on how much time I have, but I should be trying to upload this soon. I don't want to have exporting problems on the last hour of working.