Audio Added + Movement Bug
I added all audio to the game but there is a really annoying bug that came along with these changes. For some reason, when you play a level, the player is stuck walking to the right. I think this might be because of the audio slider, so if need be I can get rid of it. I don't want to do that though because it is really helpful, so I want to find a workaround.